requirements.txt
fastapi
requests
beautifulsoup4
google-generativeai
python-dotenv
SQLAlchemy
PyMySQL
uvicorn
tavily-python